factor into library
cleanup
